made some changes for income split in the related classes

+++ src/main/java/org/moneymatters/mm_backend/models/dto/TransactionDTO.java
@@ -40,7 +40,7 @@ public class TransactionDTO {
         this.tagId = tagId;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
